% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add}
\alias{add}
\title{Add the inputs together.}
\usage{
add(x, y)
}
\arguments{
\item{x}{A numeric value to be added.}

\item{y}{A numeric value to be added.}
}
\value{
The input values summed together.
}
\description{
This function takes two numbers and returns the sum.
}
