% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_determination.R
\name{coef_determination}
\alias{coef_determination}
\title{Function to compute coefficient of determination}
\usage{
coef_determination(x, y)
}
\arguments{
\item{x}{A numeric vector of length n.}

\item{y}{A numeric vector also of length n.}
}
\value{
The coefficient of determination of x and y.
}
\description{
This function takes two numbers and returns their sum.
}
\examples{
coef_determination(c(1, 2, 3, 4), c(4, 6, 7, 4))
}
