% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohens_d.R
\name{cohens_d}
\alias{cohens_d}
\title{Function to compute cohen's d.}
\usage{
cohens_d(x1, x2)
}
\arguments{
\item{x1}{A numeric vector of length n1.}

\item{x2}{A numeric vector of length n2.}
}
\value{
The cohen's d of x1 and x2.
}
\description{
This function takes two vectors and returns the absolute value of cohen's d.
}
\examples{
cohens_d(c(1, 2, 3, 4), c(4, 6, 7, 4))
}
