% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pearson_corr.R
\name{pearson_corr}
\alias{pearson_corr}
\title{Function to compute pearson's correlation coefficient}
\usage{
pearson_corr(x, y)
}
\arguments{
\item{x}{A numeric vector of length n.}

\item{y}{A numeric vector also of length n.}
}
\value{
The pearson correlation coefficient of x and y.
}
\description{
This function takes two numbers and returns their sum.
}
\examples{
pearson_corr(c(1, 2, 3, 4), c(4, 6, 7, 4))
}
