% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_ratio.R
\name{risk_ratio}
\alias{risk_ratio}
\title{Function to compute risk ratio.}
\usage{
risk_ratio(data)
}
\arguments{
\item{data}{A 2x2 numeric matrix of the form: c(Diseased_Exposed, NonDiseased_Exposed, Disease_NonExposed, NonDisease_NonExposed2)}
}
\value{
The risk ratio of the matrix.
}
\description{
This function takes a 2x2 matrix and calculates the risk ratio.
}
\examples{
risk_ratio(matrix(c(20, 6, 380, 594), nrow = 2, ncol = 2))
}
