% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds_ratio.R
\name{odds_ratio}
\alias{odds_ratio}
\title{Function to compute odds ratio with normal approximation (Wald)}
\usage{
odds_ratio(data)
}
\arguments{
\item{data}{A 2x2 numeric matrix of the form: c(Diseased_Exposed, NonDiseased_Exposed, Disease_NonExposed, NonDisease_NonExposed2)}
}
\value{
The odds ratio of the matrix.
}
\description{
This function takes a 2x2 matrix and calculates the odds ratio.
}
\examples{
odds_ratio(matrix(c(20, 6, 380, 594), nrow = 2, ncol = 2))
}
