% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pearson_corr.R
\name{pearson_corr}
\alias{pearson_corr}
\title{Function to compute pearson's correlation coefficient}
\usage{
pearson_corr(x, y)
}
\arguments{
\item{x}{A numeric vector of length n.}

\item{y}{A numeric vector also of length n.}
}
\value{
A numeric value of the pearson correlation coefficient of x and y.
}
\description{
This function takes two vectors and returns pearson's correlation coefficient. The formula for this is:
\deqn{r_{xy} = \frac{\sum_{i=1}^{n} (x_i - \bar{x})(y_i - \bar{y})}{\sqrt{\sum_{i=1}^{n} (x_i - \bar{x})^2} \sqrt{\sum_{i=1}^{n} (y_i - \bar{y})^2}}}
}
\examples{
pearson_corr(c(1, 2, 3, 4), c(4, 6, 7, 4))
# Output: 0.0860663
}
